UPDATE: add sketchy implementation for porting `ColorTextEdtior`
Note that we've added an function `get_current_context_ptr()`
to pass pointer of `ImGuiContext` since we cannot access the actual
memory address of it by the value returned by a Cython-wrapped
`get_current_context()` function. (just like the reason mentioned
in [1], a C object will be wrapped by a Python object, so that
we cannot access its real address even if we have defined a `_ptr`
field in the Cython extension class)

Besides, those functions for setting/getting context should also
be implemented in `pyimgui` (i.e. `get_current_context_ptr()` and
`set_current_context_ptr()`, but at least the former one).

And note that this might not be a good solution since we have to
expose the real address of internal object, so that these functions
are classified as workarounds and we should figure out a better
ones in the future.

[1]: https://github.com/NaleRaphael/codememo/pull/1